+++ cmd/migrate/cockroachdb-operator/migrate-operator.go
@@ -1,7 +1,8 @@
-package cockroachdb_enterprise_operator
+package cockroachdb_operator
 
 import (
 	"fmt"
+
 	"github.com/cockroachdb/helm-charts/pkg/migrate"
 	"github.com/spf13/cobra"
 )
@@ -13,16 +14,16 @@ var (
 
 var buildManifestFromOperator = &cobra.Command{
 	Use:   "operator",
-	Short: "Generate migration manifests for cockroach-enterprise-operator from cockroach-operator(https://github.com/cockroachdb/cockroach-operator)",
-	Long: `Generate the required Kubernetes manifests to assist in migrating from the CockroachDB Operator 
-to the CockroachDB Enterprise Operator.
+	Short: "Generate migration manifests for the CockroachDB Operator from the Public Operator (https://github.com/cockroachdb/cockroach-operator)",
+	Long: `Generate the required Kubernetes manifests to assist in migrating from the Public Operator
+to the CockroachDB Operator.
 
 This command is designed to simplify the migration process by generating manifests that are compatible 
-with the Cockroach Enterprise Operator, including resources such as CRDBNode and values.yaml
+with the CockroachDB Operator, including resources such as CRDBNode and values.yaml
 
-It is intended for users who initially deployed CockroachDB via the Cockroach Operator(https://github.com/cockroachdb/cockroach-operator) 
+It is intended for users who initially deployed CockroachDB via the Public Operator (https://github.com/cockroachdb/cockroach-operator)
 and now wish to take advantage of the additional capabilities and lifecycle management features offered 
-by the Cockroach Enterprise Operator.
+by the CockroachDB Operator.
 
 Always review the generated manifests thoroughly and test in a staging environment before applying changes 
 to a production cluster.
@@ -51,7 +52,7 @@ func buildManifestFromCockroachDBOperator(cmd *cobra.Command, args []string) err
 	fmt.Printf("📁 Output directory: %s\n", outputDir)
 	fmt.Println("📌 Next steps:")
 	fmt.Printf("   1. Review the generated YAML files under the '%s' directory.\n", outputDir)
-	fmt.Println("   2. Follow the README.md under scripts/migration/operator directory")
+	fmt.Println("   2. Follow the manual_migration.md under docs/migration/operator directory")
 	fmt.Println("   3. Monitor the cluster to ensure a smooth transition.")
 	fmt.Printf("\n⚠️ WARNING:\n")
 	fmt.Println("   Always review the generated manifests thoroughly and test in staging environment")